+++ web/src/main/java/com/interview/vatcalculator/validation/VatCalculatorValidator.java
@@ -1,7 +1,0 @@
-package com.interview.vatcalculator.validation;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class VatCalculatorValidator {
-}