+++ web/src/main/java/com/interview/vatcalculator/validation/VatConstraint.java
@@ -1,4 +1,0 @@
-package com.interview.vatcalculator.validation;
-
-public interface VatConstraint {
-}