+++ web/src/main/java/com/interview/vatcalculator/validation/constraint/VatRateConstraint.java
@@ -1,0 +1,20 @@
+package com.interview.vatcalculator.validation.constraint;
+
+
+import com.interview.vatcalculator.validation.validator.VatRateValidator;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.*;
+
+@Documented
+@Constraint(validatedBy = VatRateValidator.class)
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface VatRateConstraint {
+    String message() default "VAT rate is missing or invalid.";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}