+++ web/src/main/java/com/interview/vatcalculator/validation/model/Error.java
@@ -1,0 +1,20 @@
+package com.interview.vatcalculator.validation.model;
+
+public class Error {
+
+    private final String code;
+    private final String message;
+
+    public Error(String code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}