+++ web/src/main/java/com/interview/vatcalculator/web/VatCalculatorEndpoint.java
@@ -1,6 +1,6 @@
 package com.interview.vatcalculator.web;
 
-import com.interview.vatcalculator.web.api.VatCalculatorQuery;
+import com.interview.vatcalculator.web.api.VatCalculatorRequest;
 import com.interview.vatcalculator.web.api.VatCalculatorResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -10,14 +10,14 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.springdoc.api.annotations.ParameterObject;
 import org.springframework.http.MediaType;
-import org.springframework.web.ErrorResponse;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 
 public interface VatCalculatorEndpoint {
 
-    @Operation(description = "Retrieves vat amounts based on either net, vat or gross amount " +
+    @Operation(description = "Calculates amounts based on either net, vat or gross amount " +
             "and a valid Austrian VAT rate.",
             parameters = {
                     @Parameter(in = ParameterIn.QUERY, name = "net", description = "Net amount.",
@@ -30,18 +30,18 @@ public interface VatCalculatorEndpoint {
                             schema = @Schema(implementation = Double.class)),
             },
             responses = {
-                    @ApiResponse(responseCode = "200", description = "VAT amount successfully retrieved.",
+                    @ApiResponse(responseCode = "200", description = "Amounts successfully retrieved.",
                             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                                     schema = @Schema(implementation = VatCalculatorResponse.class))),
                     @ApiResponse(responseCode = "400", description = "Invalid data in request.",
                             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                                    schema = @Schema(implementation = ErrorResponse.class))),
+                                    schema = @Schema(implementation = VatCalculatorResponse.class))),
                     @ApiResponse(responseCode = "500", description = "Server side problem.",
                             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                                    schema = @Schema(implementation = ErrorResponse.class)))
+                                    schema = @Schema(implementation = VatCalculatorResponse.class)))
             })
 
     @GetMapping(path = "/calculate", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    VatCalculatorResponse getVatAmounts(@ParameterObject VatCalculatorQuery query);
+    ResponseEntity<VatCalculatorResponse> calculateAmounts(@ParameterObject VatCalculatorRequest request);
 }