+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 All release related information to vat-calculator service will be documented in this file.
 
 ## Introduction
-The service is providing capabilities to calculate vat amounts based on specific data provided.
+The service is providing capabilities to calculate VAT related amounts based on specific data provided.
 
 ## Unreleased
 ### Changed
 - API endpoint /calculate
-  - Calculate vat amounts based on either net, vat or gross amount and a valid Austrian VAT rate.
+  - Calculate VAT related amounts based on either net, vat or gross amount and a valid Austrian VAT rate.